+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [Issue #178](https://github.com/prb28/vscode-amiga-assembly/issues/178) Create ADF with defined Bootblock not work
 - [Feat #176](https://github.com/prb28/vscode-amiga-assembly/issues/176) Compress the extension with webpack
 - [Pull #173](https://github.com/prb28/vscode-amiga-assembly/pull/173) *dansalvato* Data Generator minor fixes
+- [Feat #165](https://github.com/prb28/vscode-amiga-assembly/issues/165) Memory breakpoint
 
 ### [1.0.0] - 2021-11-27
 - Configuration *checkErrorOnSave* to disable compile on save to check errors